+++ ModelSwapperSkins/BoneMapping/Bone.cs
@@ -1,23 +1,44 @@
-﻿using System;
+﻿using ModelSwapperSkins.Utils;
+using System;
 using UnityEngine;
 
 namespace ModelSwapperSkins.BoneMapping
 {
     [Serializable]
-    public class Bone
+    public class Bone : ICloneable
     {
         public BoneInfo Info;
         public Transform BoneTransform;
         public string ModelPath;
 
+        public Bone(BoneInfo info, Transform boneTransform, string modelPath)
+        {
+            Info = info;
+            BoneTransform = boneTransform;
+            ModelPath = modelPath;
+        }
+
+        public Bone(BoneInfo info, string modelPath, Transform rootTransform) : this(info, rootTransform.Find(modelPath), modelPath)
+        {
+        }
+
+        public Bone(BoneInfo info, Transform rootTransform, Transform boneTransform) : this(info, boneTransform, TransformUtils.GetObjectPath(boneTransform, rootTransform))
+        {
+        }
+
         public Bone MakeCopyFor(BonesProvider displayBonesProvider)
         {
-            return new Bone
-            {
-                Info = Info,
-                ModelPath = ModelPath,
-                BoneTransform = displayBonesProvider.transform.Find(ModelPath)
-            };
+            return new Bone(Info, ModelPath, displayBonesProvider.transform);
+        }
+
+        public Bone Clone()
+        {
+            return new Bone(Info, BoneTransform, ModelPath);
+        }
+
+        object ICloneable.Clone()
+        {
+            return Clone();
         }
     }
 }