+++ ModelSwapperSkins/BoneMapping/InitializerRules/BoneInitializerRules_Railgunner.cs
@@ -29,6 +29,16 @@ namespace ModelSwapperSkins.BoneMapping.InitializerRules
                     case BoneType.Toe1R:
                         bone.RotationOffset *= Quaternion.Euler(0f, 270f, 0f);
                         break;
+                    case BoneType.Stomach:
+                        bone.Scale *= 0.7f;
+                        break;
+                    case BoneType.Chest:
+                        bone.PositionOffset += new Vector3(0f, -0.15f, 0f);
+                        bone.Scale *= 0.7f;
+                        break;
+                    case BoneType.Head:
+                        bone.PositionOffset += new Vector3(0f, -0.125f, 0f);
+                        break;
                 }
 
                 return bone;