+++ ModelSwapperSkins/ModelParts/ModelPartsInitializer.cs
@@ -352,6 +352,10 @@ namespace ModelSwapperSkins.ModelParts
             OverrideParts("RoR2/Base/Titan/TitanGoldBody.prefab",
                           new ModelPartConstructor("TitanArmature/ROOT/base/stomach/chest/upper_arm.r/lower_arm.r/hand.r/RightFist/Sword", ModelPartFlags.Weapon));
 
+            OverrideParts("RoR2/Base/Treebot/TreebotBody.prefab",
+                          new ModelPartConstructor("TreebotFlowerMesh", ModelPartFlags.BodyFeature),
+                          new ModelPartConstructor("TreebotRootMesh", ModelPartFlags.BodyFeature));
+
             OverrideParts("RoR2/DLC1/VoidSurvivor/VoidSurvivorBody.prefab",
                           new ModelPartConstructor("mdlVoidSurvivorMetal", ModelPartFlags.BodyFeature),
                           new ModelPartConstructor("metalcollar.001", ModelPartFlags.Body));