+++ ModelSwapperSkins/Main.cs
@@ -14,7 +14,7 @@ namespace ModelSwapperSkins
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "Gorakh";
         public const string PluginName = "ModelSwapperSkins";
-        public const string PluginVersion = "1.1.0";
+        public const string PluginVersion = "1.2.0";
 
         void Awake()
         {