+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**?.?.? Changes:**
+
+* Fixed:
+  * Mul-t pelvis rotation
+
 **1.1.0 Changes:**
 
 * Added: